Modifications 1. added the cer table after stage-1 test 2. if the total cer>1 for the intersection hypothesis the adaptive boundary computation needs adjustment. attaching Martin Posh's comments for the details.
Martin.P
The problem only arises for non-parametric or mixed tests (as in the example with 4 treatments and  2 endpoints). I think the problem is that in the case that B_J is larger than the number of hypothesis with positive second stage weights that are continued to the second stage, there is no solution for the boundary within [0,1]. Even if all boundaries are equal to 1, then the total conditional error of the second stage test is  equal to the number of hypotheses continued to the second stage, and thus lower than B_J.  If this is the case, one should just reject H_J (because then I could choose C_J=Infinity) and the second stage total conditional error is still smaller than B_J). Therefor the boundary should be 1 one in this case (and not 0?).

In my code I just check the condition if B_J >= # Hypotheses with pos. second stage weights, continued to the second stage. If it is met I reject H_J. If not I compute  c_J with unroot.

I encountered this issue in the setting with 2 Endpoints in the paper, for example for the ultra selection rule, when there is a treatment effect in all treatments and endpoints. It can happen then that sum of conditional errors becomes larger than one - but since only one treatment is continued to the second stage and endpoints are tested hierarchically, such that only one hypothesis has positive weight, no c_J can be found that leads to a total CER >1.